requirements.txt
django
djangorestframework
django-cors-headers
channels
channels-redis
psycopg2-binary
django-redis
Pillow
